Disable array and string re-embedding with MMTk.
The current array-reembedding code does not forward the references in
the out-of-object buffer, and will result in dangling references.  And
the proper place to implement array or string re-embedding is in
`ObjectModel::copy` in the Rust part of the binding.  So we disable the
code path to re-embedding for now.