Fix global ivtbl handling
The upstream now updates global ivtbl entries when updating weak tables.
While this works well for the default mark-and-compact GC, it is not
idiomatic to separate the marking and the updating unless we are
implementing a compacting GC.

We restored the previous eager-updating strategy when using MMTk.  For
this, we resurrected `rb_mv_generic_ivar`, too.