Remove the HAS_MOVED_GIVTBL bit
We no longer use the HAS_MOVED_GIVTBL bit in the hidden header.  Now
when we try to get the givtbl of a potentially moved object, we first
look up the global general_iv_tbl_ and, if not found, look up the
RubyBinding::moved_givtbl table in the Rust part of the binding.