Don't acquire VM lock to get gen_ivtbl in MMTk.
During GC, MMTk worker threads may need to get the original gen_ivtbl of
an object when the object is moved.  It crashed when a GC worker
attempts to acquire VM lock when there are multiple ractors.  We skip
the VM lock when getting gen_ivtbl from a MMTk GC worker thread.