Unconditionally increase imemo type size
We previously used conditional compilation to increase the flag bits for
imemo types from 4 to 5 in order to accommodate imemo:mmtk_strbuf and
imemo:mmtk_objbuf when using MMTk.  It changed the value of
ISEQ_TRANSLATED which YJIT hardcoded.  Since it is difficult to do
conditional compilation for YJIT and ZJIT, we change the number iseq
type bits and the values of IMEMO_FL_USERx unconditionally.