Handle cc_refinement table and reorganize code
Previously we didn't process the newly added cc_refinement table, and
caused crashes.

We reorganized the code for handling weak references when using MMTk.
We moved most of the code to mmtk_support.c, and follow the order as
listed in `rb_gc_vm_weak_table_foreach` in `gc.c`, so that we don't miss
a table.