Fix forking in multi-ractor scenario
When there are multiple Ractors, there is a chance that one Ractor
triggers GC when another Ractor has just stopped all MMTk GC work
threads in order to fork.  If this happens, the forking thread will be
blocking waiting for the Ractor sched barrier, and the Ractor that
triggered GC will be waiting for GC threads to finish GC, which will
never happen.

We move the invocations of `rb_gc_before_fork()` and
`rb_gc_after_fork()` inside the `RB_VM_LOCKING() { ... }` block in
`rb_fork_ruby`.  By doing this, either the forking Ractor will wait for
GC to finish before stopping GC worker threads, or the Ractor that
triggers GC cannot stop other threads until the forking ractor respawned
GC worker threads.

We also hold the VM lock when running finalizers when the VM is about to
exits.