Use square brackets in RUBY_DESCRIPTION
`EnvUtil::current_parser` uses a regexp to detect the presence of
`+PRISM`, but it is not expecting round brackets in the string.  We
workaround this by changing round brackets to square brackets.

We no longer need to exclude the test
`TestPatternMatching::test_alternative_pattern_nested` because it works
properly now with PRISM correctly detected.